requirements.txt
Appium-Python-Client
selenium
pytest
pytest-html